PR feedback iterations. also dedupe sample on checksum: if exists, return existing.
