+++ dev.js
@@ -3,12 +3,12 @@ const chokidar = require('chokidar')
 
 electron.start()
 
-chokidar.watch('main.js', {
-  ignored: '*/node_modules/*'
-})
-.on('change', () => electron.restart())
+chokidar
+  .watch('main.js')
+  .on('change', () => electron.restart())
 
-chokidar.watch('**/*.(js|json|html)', {
-  ignored: '*/node_modules/*'
-})
-.on('change', () => electron.reload())
+chokidar
+  .watch('**/*.(js|json|html)', {
+    ignored: '**/data/**'
+  })
+  .on('change', () => electron.reload())
