+++ package.json
@@ -29,7 +29,6 @@
     "babel-preset-react": "^6.16.0",
     "babel-register": "^6.18.0",
     "core-js": "^2.4.1",
-    "crypto": "0.0.3",
     "debug": "^2.3.3",
     "electron": "^1.4.7",
     "globals": "^9.14.0",
