+++ samples/actions.js
@@ -11,6 +11,7 @@ module.exports = createActions(
   'LOAD_SAMPLE_END',
   'CREATE_SAMPLE',
   'CREATE_SAMPLE_SUCCESS',
+  'CREATE_SAMPLE_DUPLICATE',
   'CREATE_SAMPLE_FAILURE',
   'CREATE_SAMPLE_END',
   'ANALYZE_SAMPLE',
